+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 - Added a `--redirect` option to `bin/polymer-bundler`.
 - Added a `--root` option to `bin/polymer-bundler`.
+- The `stripComments` behavior now removes comments inside templates.
 <!-- Add new, unreleased changes here. -->
 
 ## 2.0.3 - 2017-06-07