+++ CHANGELOG.md
@@ -7,7 +7,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 - Fixed an error when generating source maps
+- Added a `--redirect` option to `bin/polymer-bundler`.
+- Added a `--root` option to `bin/polymer-bundler`.
+- The `stripComments` behavior now removes comments inside templates.
 <!-- Add new, unreleased changes here. -->
+
+## 2.0.3 - 2017-06-07
+- Fixed issue with lazy-imports impacting bundle inlining logic. https://github.com/Polymer/polymer-bundler/issues/536
+- Fixed issue with hidden div being created inside dom-modules. https://github.com/Polymer/polymer-bundler/issues/535
+- Fixed issue with injected dependencies inside dom-modules in shell when using lazy-imports. https://github.com/Polymer/polymer-bundler/issues/534
+
+## 2.0.2 - 2017-06-02
+- In cases such as the shell merge strategy, imports which are not originally
+  part of the shell are now inserted before other imports which are dependent
+  on them.  Fixes https://github.com/Polymer/polymer-bundler/issues/517.
 
 ## 2.0.1 - 2017-05-19
 - Don't include the `by-polymer-bundler` hidden div in bundled files